Update app.js v6.11
changed:       direction = -st.pigStartDirection;

At the end of cake eating so that the pig continues in the opposite way from when the action started. This is because we have the pig continue the movement until it turns around, therefore we need the reversal after it is done eating.
Also undid fix in 6.10 for this.